+++ packages/web-sdk/src/instrumentations/performance/performanceUtils.ts
@@ -30,6 +30,7 @@ export function calculateFaroResourceTiming(resourceEntryRaw: PerformanceResourc
     decodedBodySize,
     domainLookupEnd,
     domainLookupStart,
+    duration,
     encodedBodySize,
     fetchStart,
     initiatorType,
@@ -51,6 +52,7 @@ export function calculateFaroResourceTiming(resourceEntryRaw: PerformanceResourc
 
   return {
     name: name,
+    duration: toFaroPerformanceTimingString(duration),
     tcpHandshakeTime: toFaroPerformanceTimingString(connectEnd - connectStart),
     dnsLookupTime: toFaroPerformanceTimingString(domainLookupEnd - domainLookupStart),
     tlsNegotiationTime: toFaroPerformanceTimingString(requestStart - secureConnectionStart),
@@ -96,7 +98,6 @@ export function calculateFaroNavigationTiming(navigationEntryRaw: PerformanceNav
     domContentLoadedEventEnd,
     domContentLoadedEventStart,
     domInteractive,
-    duration,
     fetchStart,
     loadEventEnd,
     loadEventStart,
@@ -106,7 +107,6 @@ export function calculateFaroNavigationTiming(navigationEntryRaw: PerformanceNav
 
   return {
     visibilityState: document.visibilityState,
-    totalNavigationTime: toFaroPerformanceTimingString(duration),
     pageLoadTime: toFaroPerformanceTimingString(domComplete - fetchStart),
     domProcessingTime: toFaroPerformanceTimingString(domComplete - domInteractive),
     domContentLoadHandlerTime: toFaroPerformanceTimingString(domContentLoadedEventEnd - domContentLoadedEventStart),
@@ -117,6 +117,8 @@ export function calculateFaroNavigationTiming(navigationEntryRaw: PerformanceNav
     ttfb: toFaroPerformanceTimingString(Math.max(responseStart - (activationStart ?? 0), 0)),
 
     type: type,
+
+    ...calculateFaroResourceTiming(navigationEntryRaw),
   };
 }