+++ packages/web-tracing/src/faroTraceExporter.ts
@@ -3,6 +3,7 @@ import type { ExportResult } from '@opentelemetry/core';
 import { createExportTraceServiceRequest } from '@opentelemetry/otlp-transformer';
 import type { ReadableSpan, SpanExporter } from '@opentelemetry/sdk-trace-web';
 
+import { sendFaroEvents } from './faroTraceExporter.utils';
 import type { FaroTraceExporterConfig } from './types';
 
 export class FaroTraceExporter implements SpanExporter {
@@ -12,6 +13,7 @@ export class FaroTraceExporter implements SpanExporter {
     const traceEvent = createExportTraceServiceRequest(spans, { useHex: true, useLongBits: false });
 
     this.config.api.pushTraces(traceEvent);
+    sendFaroEvents(traceEvent.resourceSpans);
 
     resultCallback({ code: ExportResultCode.SUCCESS });
   }